requirements.txt
opencv-python
opencv-python
numpy
natsort
ultralytics
pillow